estructuras separadas y agregadas a carpeta components
